CLEANUP: first attempt at splitting couchapps
I could not get couchapp pushapps to work correctly,
so we have to use the small bash script pushapp.bash.

I could not get the lib directory to just be a symlink
of a directory in common (couchapp wouldn't push it), so
it is a copy. Same for vendor, style and index.html.

Note that this means that any modification to the copies
will be lost, only edit files in common !