EVENTLY: support changes filters.
evently will open one longpoll connection per db per filter. To
use filters, use a _changes event with a field called filter.
This field should contain the following json object:
{
  "name": ..
  "ddoc": ..
}
If ddoc is not present, the default value of the design document
serving the current page is used.
Name is the name of the filter function